Add files via upload
cleaned up the code a bit from finalProject.py. 
TODO: currently code is crashing on the 'meanv' function (function that finds the mean vector given a list of vectors).
I suspect this is because the input list (during the algorithm process) is a list of vectors of possibly different dimensions. I made note of this in the appropriate locations in code.