Update finalProject.py
solved problem with program malfunctioning when there existed a cluster with 0 datapoints in it. 

New problem: are we defining distance correctly? we are assigning data points to be under a certain cluster if they are "closest" to that cluster via l2 norm. But "closest" here does not mean "most similar in meaning"! This would require cosine metric and NOT l2 norm!